W-12967361: Handle spllit package issue between mule-module-extensions-spring-support and mule-module-extensions-support, remove AbstractResourcesGenerator
